change format to viewer
