+++ packages/frontend/package.json
@@ -30,7 +30,7 @@
     "prop-types": "15.8.1",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "react-firebase-hooks": "^5.1.1",
+    "react-firebase-hooks": "5.1.1",
     "react-hook-form": "7.44.3",
     "ts-toolbelt": "9.6.0",
     "typescript": "5.1.3",
