--- packages/frontend/src/app/home/layout.tsx
@@ -1,14 +1,0 @@
-import { AuthProvider } from "@/hooks/login";
-
-export default function HomeLayout({
-    children,
-  }: {
-    children: React.ReactNode;
-  }) {
-    return (
-      <AuthProvider>
-        <main >{children}</main>
-      </AuthProvider>
-    );
-  }
-
