+++ packages/frontend/src/app/login/layout.tsx
@@ -1,0 +1,7 @@
+export default function LoginLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return <main>{children}</main>;
+}
