+++ packages/frontend/src/hooks/index.tsx
@@ -1,3 +1,4 @@
 export * from "./useToastAlertProvider";
 export * from "./useGeneratedQuery";
 export * from "./useGeneratedMutation";
+export * from "./login";
